Allow warwick-legend to run in G4 interactive mode
Update main program following Geant4 examples to create and use an
interactive session when no macro file is supplied. This will
launch the "best" UI/Vis driver available from the used Geant4 install.
In most cases this will be Qt or tcsh+OpenGL/X11.

Provide a simple test to run warwick-lebgend in batch mode to assure
this is not affected.