Fixes for errors when switching geometries
When constructing geometry, manually clear stores. Whilst
ReinitializeGeometry ensures this is called on the next run, it
cannot automatically clear the stores without causing a segfault.

Because sensitive detector persists, store in a thread-local G4Cache
wrapper, only constructing and adding the the SD manager once. This
still appears to show erros when run through ASAN, but further checks
needed as it follows the Geant4 examples using the same pattern.

Retain and simplify biasing setup in ConstructSDandField, but ownership
still requires checking.