Enable switching between baseline and alternative geometries (#19)
* Change geometry at runtime via macro command

WLGD currently provides two geometry setups ("baseline" and
"alternative"), but the command to swap between them was reported
to not work.

Use method-based generic messenger to bind new "setGeometry" command to
a member function of the detector construction that swaps the geometry.
Allow command to take "baseline" and "alternative" strings as only two
allowed arguments. Follow Geant4 examples for kernel function to call
to reinit geometry.

Add basic runtime test to exercise swapping of geometry.

* Update .clang-tidy exclusions

* Fixes for errors when switching geometries

When constructing geometry, manually clear stores. Whilst
ReinitializeGeometry ensures this is called on the next run, it
cannot automatically clear the stores without causing a segfault.

Because sensitive detector persists, store in a thread-local G4Cache
wrapper, only constructing and adding the the SD manager once. This
still appears to show erros when run through ASAN, but further checks
needed as it follows the Geant4 examples using the same pattern.

Retain and simplify biasing setup in ConstructSDandField, but ownership
still requires checking.

* Mount /cvmfs in container per docs

Seemingly needed to ensure data files can be read, but not clear if required.