Do not store trajectories by default
Initial runs showed large and increasing memory consumption by the
application. Runs with valgrind could not pinpoint anything specific
in WLGD code. Further investigation showed that memory increases occured
with trajectory storage, likely due to the large number of trajectories
that can potentially occur in a biased run, plus the use of
G4Allocator's memory pool to manage this memory.

Only add a trajectory to a track if the user has requested storage via
/tracking/storeTrajectories. Update EndOfEventAction to check for a
valid pointer to the trajectory container before iterating over it.

This will lead to branches with null data in the output TTree, but this
should be safe.