Provide UI command to export geometry to GDML file
New member function in WLGDDetectorConstruction exports current, full
geometry to a GDML output file. Bind this function to a UI command
that takes the output file path as an argument.

Implement tests that run the command and then check that the output
file exists and is non-zero length. No validation of the output is
performed as yet.